Fix Vite HMR origin/port mismatch; make fluid font-size discovery generic
vite.config.js hardcoded server.origin to port 5273 while server.port
resolved from VITE_PRIMARY_PORT (no fallback tie between the two). Any
project that sets a custom VITE_PRIMARY_PORT gets an HMR client that
tries to connect on a port nothing is listening on, so file changes
never reach the browser and the dev server looks broken even though
it's serving correct content. origin now derives from the same
VITE_PORT constant as port.

fluid-font-calculations.cjs hardcoded a flat FLUID_TOKENS list and
always emitted --fluid-text-{token}. Replaced with discoverFluidTokens(),
which checks the shared SIZE_SUFFIXES scale with no prefix (body) and
once per SUPPORTED_FILTERS prefix (e.g. headline-lg), keeping only the
combinations that have a --fs-{token}-min/max range actually defined in
font-sizes.css. propForToken() picks --fluid-text-* vs --fluid-{token}
based on whether the token is prefixed. A project can now add a
headline/ui scale (or new sizes) purely in CSS with no plugin changes.
Verified this produces byte-for-byte identical output to the old
hardcoded list for the starter's current font-sizes.css.

SIZE_SUFFIXES/SUPPORTED_FILTERS/fontNames/fontSlugs moved into a new
shared src/theme-json/helpers/font-scale.cjs so fonts.js and the plugin
source from one place instead of maintaining separate copies. Plain
CommonJS so the plugin can require() it directly and fonts.js (ESM) can
import its named exports via Node's standard CJS interop, without the
two files circularly requiring each other.
