+++ wp-content/themes/wp-starter/src/plugins/fluid-font-calculations.cjs
@@ -9,10 +9,46 @@
 
 const fs = require('fs');
 const path = require('path');
+const { SIZE_SUFFIXES, SUPPORTED_FILTERS } = require('../theme-json/helpers/font-scale.cjs');
 
 const THEME_ROOT = path.join(__dirname, '..', '..');
 
-const FLUID_TOKENS = ['sm', 'base', 'md', 'lg', 'xl', '2xl'];
+/**
+ * The output custom property for a token: filter-prefixed tokens (e.g. `headline-lg`) get
+ * `--fluid-{token}`, everything else (the body scale) gets `--fluid-text-{token}`. Suffixes in
+ * SIZE_SUFFIXES never contain a hyphen, so any hyphen in the token means it's filter-prefixed.
+ *
+ * @param {string} token
+ */
+function propForToken(token) {
+	return token.includes('-') ? `--fluid-${token}` : `--fluid-text-${token}`;
+}
+
+/**
+ * Checks SIZE_SUFFIXES once with no prefix (the body scale), then once per SUPPORTED_FILTERS
+ * prefix (e.g. `headline-lg`), keeping only the combinations that have a --fs-{token}-min/max
+ * range actually defined in font-sizes.css.
+ *
+ * @param {string} src font-sizes.css content
+ * @returns {string[]}
+ */
+function discoverFluidTokens(src) {
+	const hasRange = (token) =>
+		new RegExp(`--fs-${token}-min:`).test(src) && new RegExp(`--fs-${token}-max:`).test(src);
+
+	const tokens = SIZE_SUFFIXES.filter(hasRange);
+
+	for (const filter of SUPPORTED_FILTERS) {
+		for (const suffix of SIZE_SUFFIXES) {
+			const token = `${filter}-${suffix}`;
+			if (hasRange(token)) {
+				tokens.push(token);
+			}
+		}
+	}
+
+	return tokens;
+}
 
 /** @param {string} themeRoot */
 function readFontSizes(themeRoot) {
@@ -23,18 +59,14 @@ function readFontSizes(themeRoot) {
 		throw new Error('[fluid-font-calculations] Could not parse --cfg-fluid-bp-min from font-sizes.css');
 	}
 	const bpMin = parseInt(bpMinM[1], 10);
+	const tokens = discoverFluidTokens(src);
 	const ranges = {};
-	for (const token of FLUID_TOKENS) {
-		const minRe = new RegExp(`--fs-${token}-min:\\s*([^;]+);`, 'm');
-		const maxRe = new RegExp(`--fs-${token}-max:\\s*([^;]+);`, 'm');
-		const minM = src.match(minRe);
-		const maxM = src.match(maxRe);
-		if (!minM || !maxM) {
-			throw new Error(`[fluid-font-calculations] Missing --fs-${token}-min/max in font-sizes.css`);
-		}
+	for (const token of tokens) {
+		const minM = src.match(new RegExp(`--fs-${token}-min:\\s*([^;]+);`, 'm'));
+		const maxM = src.match(new RegExp(`--fs-${token}-max:\\s*([^;]+);`, 'm'));
 		ranges[token] = { min: minM[1].trim(), max: maxM[1].trim() };
 	}
-	return { bpMin, ranges };
+	return { bpMin, tokens, ranges };
 }
 
 /** @param {string} themeRoot */
@@ -75,11 +107,11 @@ function buildFluidTextReplacements(opts) {
 	if (bpMax <= bpMin) {
 		throw new Error(`[fluid-font-calculations] bpMax (${bpMax}) must be > bpMin (${bpMin})`);
 	}
-	const { ranges } = readFontSizes(themeRoot);
+	const { tokens, ranges } = readFontSizes(themeRoot);
 	const rangePx = bpMax - bpMin;
 	const map = new Map();
 
-	for (const token of FLUID_TOKENS) {
+	for (const token of tokens) {
 		const { min, max } = ranges[token];
 		const minPx = lengthToPx(min, rootPx);
 		const maxPx = lengthToPx(max, rootPx);
@@ -91,16 +123,15 @@ function buildFluidTextReplacements(opts) {
 			deltaF === 0
 				? min
 				: `calc(${intercept}px + ${slopeVw}vw)`;
-		const prop = `--fluid-text-${token}`;
 		const clampVal = `clamp(var(--fs-${token}-min), ${mid}, var(--fs-${token}-max))`;
-		map.set(prop, clampVal);
+		map.set(propForToken(token), clampVal);
 	}
 	return map;
 }
 
-/** @param {string} varRef e.g. var(--fluid-text-lg) */
+/** @param {string} varRef e.g. var(--fluid-text-lg) or var(--fluid-headline-lg) */
 function tokenFromFluidVar(varRef) {
-	const m = String(varRef).trim().match(/^var\(\s*--fluid-text-(\w+)\s*\)$/i);
+	const m = String(varRef).trim().match(/^var\(\s*(--fluid-[\w-]+)\s*\)$/i);
 	return m ? m[1] : null;
 }
 
@@ -136,7 +167,9 @@ postcssPlugin.postcss = true;
 
 module.exports = {
 	THEME_ROOT,
-	FLUID_TOKENS,
+	SIZE_SUFFIXES,
+	SUPPORTED_FILTERS,
+	discoverFluidTokens,
 	readFontSizes,
 	readBreakpointContainerPx,
 	buildFluidTextReplacements,
